extract panic_hook
