+++ Cargo.toml
@@ -81,6 +81,7 @@ derive_more = "0.11.0"
 num = "0.2.0"
 hex = "0.3"
 rust-crypto = "^0.2"
+panic_hook = { path = "components/panic_hook" }
 
 [replace]
 "raft:0.3.1" = { git = "https://github.com/pingcap/raft-rs.git" }
@@ -130,6 +131,7 @@ members = [
   "components/test_coprocessor",
   "components/test_util",
   "components/codec",
+  "components/panic_hook",
 ]
 
 [profile.dev]
