+++ components/codec/Cargo.toml
@@ -8,6 +8,7 @@ path = "lib.rs"
 
 [dependencies]
 quick-error = "1.2"
+panic_hook = { path = "../panic_hook" }
 
 [dev-dependencies]
 rand = "0.5"
