+++ src/lib.rs
@@ -113,6 +113,8 @@ extern crate zipf;
 #[macro_use]
 extern crate derive_more;
 
+extern crate panic_hook;
+
 #[macro_use]
 pub mod util;
 pub mod config;
