+++ src/raftstore/store/peer_storage.rs
@@ -2115,7 +2115,7 @@ mod test {
         s.snap_state = RefCell::new(SnapState::Applying(Arc::new(AtomicUsize::new(
             JOB_STATUS_FAILED,
         ))));
-        let res = recover_safe!(|| s.cancel_applying_snap());
+        let res = ::panic_hook::recover_safe(|| s.cancel_applying_snap());
         assert!(res.is_err());
     }
 
@@ -2164,7 +2164,7 @@ mod test {
         s.snap_state = RefCell::new(SnapState::Applying(Arc::new(AtomicUsize::new(
             JOB_STATUS_FAILED,
         ))));
-        let res = recover_safe!(|| s.check_applying_snap());
+        let res = ::panic_hook::recover_safe(|| s.check_applying_snap());
         assert!(res.is_err());
     }
